Give some parameters default values when adding the API

+++ src/main/java/com/sms/courier/entity/api/common/ResponseTimeVerification.java
@@ -15,5 +15,5 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 public class ResponseTimeVerification extends BaseVerification {
 
-    private Integer timeoutLimit;
+    private long timeoutLimit;
 }
